- **Removed ArmMovementHandler**: The method onArnMove is not in the Robot Class
