+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/EPIC/Robot/Robot.java
@@ -1,5 +1,6 @@
 package org.firstinspires.ftc.teamcode.EPIC.Robot;
 
+
 import org.firstinspires.ftc.robotcore.external.Telemetry;
 import org.firstinspires.ftc.teamcode.EPIC.Components.Claw;
 import org.firstinspires.ftc.teamcode.EPIC.Components.Slider;
@@ -11,14 +12,17 @@ import org.firstinspires.ftc.teamcode.EPIC.EventListeners.IClawListener;
 import org.firstinspires.ftc.teamcode.EPIC.EventListeners.IColorListener;
 import org.firstinspires.ftc.teamcode.EPIC.EventListeners.ITouchListener;
 import org.firstinspires.ftc.teamcode.EPIC.EventListeners.TouchEventObject;
+import org.firstinspires.ftc.teamcode.EPIC.EventListeners.ArmEventObject; // Import ArmEventObject
+import org.firstinspires.ftc.teamcode.EPIC.EventListeners.IArmListener; // Import IArmListener
 import org.firstinspires.ftc.teamcode.EPIC.Motion.Mecanum_Wheels;
+import org.firstinspires.ftc.teamcode.EPIC.RobotStates.ArmStates;
 import org.firstinspires.ftc.teamcode.EPIC.Sensors.MyColorRangeSensor;
 import org.firstinspires.ftc.teamcode.EPIC.Sensors.MyTouchSensor;
 
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.hardware.HardwareMap;
 
-public class Robot implements IColorListener, ITouchListener, IClawListener {
+public class Robot implements IColorListener, ITouchListener, IClawListener, IArmListener {
     public Claw odysseyClaw;
     public Slider odysseySlider;
     public Arm odysseyArm;
@@ -32,7 +36,7 @@ public class Robot implements IColorListener, ITouchListener, IClawListener {
     private String alliance = "";
     public MyColorRangeSensor colorSensor;
 
-    public Robot(LinearOpMode parent, String alliance){
+    public Robot(LinearOpMode parent, String alliance) {
         odysseyClaw = new Claw(parent.hardwareMap);
         odysseySlider = new Slider(parent.hardwareMap);
         odysseyArm = new Arm(parent.hardwareMap);
@@ -51,19 +55,12 @@ public class Robot implements IColorListener, ITouchListener, IClawListener {
         this.parent = parent;
         this.telemetry = parent.telemetry;
         this.alliance = alliance;
-//        colorSensor = new MyColorRangeSensor(parent.hardwareMap,alliance);
-//        colorSensor.parent = parent;
-//        colorSensor.telemetry = parent.telemetry;
-//        colorSensor.addColorListener(this);
-//        touchSensor = new MyTouchSensor(parent.hardwareMap);
-//        touchSensor.initialize();
-//        touchSensor.parent = parent;
-//        touchSensor.telemetry = parent.telemetry;
-//        touchSensor.addTouchListener(this);
-//        odysseyClaw.addClawListener(this);
+
+        // Register this Robot instance as an Arm listener
+        odysseyArm.addArmListener(this); // Add this line
     }
 
-    public void setIsAutonomous(boolean isAutonomous){
+    public void setIsAutonomous(boolean isAutonomous) {
         this.isAutonomous = isAutonomous;
         odysseyClaw.setIsAutonomous(isAutonomous);
         odysseySlider.setIsAutonomous(isAutonomous);
@@ -82,70 +79,79 @@ public class Robot implements IColorListener, ITouchListener, IClawListener {
 
     @Override
     public void colorPicker(ColorEventObject event) {
-        if(this.parent.opModeIsActive()) {
+        if (this.parent.opModeIsActive()) {
             double distance = event.getDistance();
             telemetry.addData("distance", distance);
-
             telemetry.addData("color", event.getColor());
             telemetry.update();
             parent.sleep(1000);
-            //odysseyWheels.encoderDrive(0.6,distance,distance,distance,distance,1);
+            // odysseyWheels.encoderDrive(0.6, distance, distance, distance, distance, 1);
             parent.sleep(2000);
         }
     }
 
     @Override
     public void touchClicked(TouchEventObject event) {
-        if(this.parent.opModeIsActive()) {
+        if (this.parent.opModeIsActive()) {
             Thread tc = new Thread() {
                 public void run() {
-
                     if (event.getButtonStatus()) {
-                        //odysseyWheels.move(0.6, 0, 0, 0);
+                        // odysseyWheels.move(0.6, 0, 0, 0);
                     } else {
-                        //odysseyWheels.move(0, 0, 0, 0);
+                        // odysseyWheels.move(0, 0, 0, 0);
                     }
                 }
             };
             tc.start();
-
         }
     }
 
     @Override
     public void openClaw(ClawEventObject event) {
-        if(this.parent.opModeIsActive()) {
+        if (this.parent.opModeIsActive()) {
             Thread tc = new Thread() {
                 public void run() {
-
-
+                    // Add code to open the claw if needed
                 }
             };
             tc.start();
-
         }
-
     }
 
     @Override
     public void closeClaw(ClawEventObject event) {
-        if(this.parent.opModeIsActive()) {
+        if (this.parent.opModeIsActive()) {
             Thread tc = new Thread() {
                 public void run() {
-
-                    //odysseyWheels.move(0.6, 0, 0, 0);
+                    // odysseyWheels.move(0.6, 0, 0, 0);
                     try {
                         Thread.sleep(1000);
                     } catch (InterruptedException e) {
                         throw new RuntimeException(e);
                     }
-//                    odysseyWheels.move(0, 0, 0, 0);
-
+                    // odysseyWheels.move(0, 0, 0, 0);
                 }
             };
             tc.start();
-
         }
+    }
 
+    @Override
+    public void onArmMove(ArmEventObject event) {
+        ArmStates newState = event.getNewState();
+        switch (newState) {
+            case LOWERED:
+                System.out.println("Arm is in lowered position, ready to grab sample.");
+                break;
+            case LOWERED_BACK:
+                System.out.println("Arm is in lowered back position, ready to deposit.");
+                break;
+            case NEUTRAL:
+                System.out.println("Arm is back to neutral after depositing.");
+                break;
+            default:
+                System.out.println("Arm moved to an unknown state.");
+                break;
+        }
     }
 }
